Prepare For Merge of Dashboard

+++ Back-end/app/Http/Controllers/ProductController.php
@@ -11,6 +11,7 @@ use App\Models\Offer;
 use App\Models\Cart;
 use App\Models\Order;
 use App\Models\User_Order;
+
 use Illuminate\Support\Facades\DB;
 
 
@@ -42,7 +43,57 @@ class ProductController extends Controller
             return ProductResource::collection(Product::where($filterItems)->get());
         }
     }
-
+     public function catProducts(){
+        $products=Product::all();
+        $noOfAllProducts=$counterCat1=$counterCat2=$counterCat3=$counterCat4=$counterCat5=$counterCat6=$counterCat7=$counterCat8=$YCat1=$yCat2=$yCat3=$yCat4=$yCat5=$yCat6=$yCat7=$yCat8=0;
+        $AllproductsCategory=array();
+        for ($index=0; $index < count($products); $index++) { 
+            # code...
+                  if($products[$index]['category']=='lip balms'){$counterCat1++;}
+                  if($products[$index]['category']=='face care'){$counterCat2++;}
+                  if($products[$index]['category']=='deodorants'){$counterCat3++;}
+                  if($products[$index]['category']=='sun care'){$counterCat4++;}
+                  if($products[$index]['category']=='express masks'){$counterCat5++;}
+                  if($products[$index]['category']=='face radiance'){$counterCat6++;}
+                  if($products[$index]['category']=='intimate hygiene'){$counterCat7++;}
+                  if($products[$index]['category']=='skin treatments'){$counterCat8++;}
+          }
+         
+          $noOfAllProducts=$counterCat1+$counterCat2+$counterCat3+$counterCat4+$counterCat5+$counterCat6+$counterCat7+$counterCat8;
+          $YCat1=($counterCat1/$noOfAllProducts)*100;
+          $yCat2=($counterCat2/$noOfAllProducts)*100;
+          $yCat3=($counterCat3/$noOfAllProducts)*100;
+          $yCat4=($counterCat4/$noOfAllProducts)*100;
+          $yCat5=($counterCat5/$noOfAllProducts)*100;
+          $yCat6=($counterCat6/$noOfAllProducts)*100;
+          $yCat7=($counterCat7/$noOfAllProducts)*100;
+          $yCat8=($counterCat8/$noOfAllProducts)*100;
+          $data['name']='lip balms';
+          $data['y']=$YCat1;
+          array_push($AllproductsCategory,(object)$data);
+          $data['name']='face care';
+          $data['y']=$yCat2;
+          array_push($AllproductsCategory,(object)$data);
+          $data['name']='deodorants';
+          $data['y']=$yCat3;
+          array_push($AllproductsCategory,(object)$data);
+          $data['name']='sun care';
+          $data['y']=$yCat4;
+          array_push($AllproductsCategory,(object)$data);
+          $data['name']='express masks';
+          $data['y']=$yCat5;
+          array_push($AllproductsCategory,(object)$data);
+          $data['name']='face radiance';
+          $data['y']=$yCat6;
+          array_push($AllproductsCategory,(object)$data);
+          $data['name']='intimate hygiene';
+          $data['y']=$yCat7;
+          array_push($AllproductsCategory,(object)$data);
+          $data['name']='skin treatments';
+          $data['y']=$yCat8;
+          array_push($AllproductsCategory,(object)$data);
+          return $AllproductsCategory;
+     }
     public function show($id)
     {
         $product = Product::find($id);
@@ -87,15 +138,23 @@ class ProductController extends Controller
     }
     public function destroy($id)
     {
+       
         $product=Product::find($id);
+       
         $path=public_path('public/').$product->image;
         @unlink($path);
-        Offer::where('productId',$id)->delete();
-        // Cart::where('product_id',$id)->delete();
-        // $order=Order::where('product_id',$id);
-        // User_Order::where('order_id',$order->id);
-
-        // $order->delete();
+      
+        if(Offer::where('productId',$id)){
+            Offer::where('productId',$id)->delete();
+        }  
+        if(Cart::where('product_id',$id)){
+            Cart::where('product_id',$id)->delete();
+        }  
+        
+       if(Order::where('product_id',$id)->get()){  
+          Order::where('product_id',$id)->delete();
+       }
+      
          $product->delete();
     }
 }
